Migrate top level navigation to shared common
