+++ sample/android/src/main/kotlin/com/flaringapp/compose/topbar/sample/android/MainActivity.kt
@@ -18,102 +18,16 @@ package com.flaringapp.compose.topbar.sample.android
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.BackHandler
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.scaleIn
-import androidx.compose.animation.scaleOut
-import androidx.compose.animation.togetherWith
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.runtime.snapshots.SnapshotStateList
-import androidx.compose.runtime.toMutableStateList
-import androidx.compose.ui.Modifier
-import com.flaringapp.compose.topbar.sample.shared.ui.samples.CollapsingTopBarSample
-import com.flaringapp.compose.topbar.sample.shared.ui.samples.CollapsingTopBarSampleGroups
-import com.flaringapp.compose.topbar.sample.shared.ui.samples.gallery.SamplesGallery
-import com.flaringapp.compose.topbar.sample.shared.ui.samples.gallery.SamplesGalleryGroup
-import com.flaringapp.compose.topbar.sample.shared.ui.theme.ComposeCollapsingTopBarTheme
+import com.flaringapp.compose.topbar.sample.shared.App
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            ComposeCollapsingTopBarTheme {
-                SamplesNavigation()
-            }
+            App()
         }
-    }
-}
-
-@Composable
-private fun SamplesNavigation() {
-    var selectedSample: CollapsingTopBarSample? by remember {
-        mutableStateOf(null)
-    }
-
-    BackHandler(
-        enabled = selectedSample != null,
-    ) {
-        selectedSample = null
-    }
-
-    AnimatedContent(
-        modifier = Modifier.fillMaxSize(),
-        label = "SampleContentAnimation",
-        targetState = selectedSample,
-        transitionSpec = {
-            val isOpeningSample = targetState != null
-            if (isOpeningSample) {
-                fadeIn() + scaleIn(initialScale = 1.1f) togetherWith fadeOut() using null
-            } else {
-                fadeIn() togetherWith fadeOut() + scaleOut(targetScale = 1.1f) using null
-            }
-        },
-    ) { currentSelectedSample ->
-        if (currentSelectedSample != null) {
-            currentSelectedSample.Content(
-                onBack = { selectedSample = null },
-            )
-            return@AnimatedContent
-        }
-
-        SamplesGallery(
-            groups = rememberSampleGroups(),
-            onSampleSelect = { selectedSample = it },
-        )
-    }
-}
-
-@Composable
-private fun rememberSampleGroups(): SnapshotStateList<SamplesGalleryGroup> {
-    return remember {
-        mutableStateListOf(
-            SamplesGalleryGroup(
-                name = "Basic Samples",
-                samples = CollapsingTopBarSampleGroups.Basic.toMutableStateList(),
-            ),
-            SamplesGalleryGroup(
-                name = "Column Samples",
-                samples = CollapsingTopBarSampleGroups.Column.toMutableStateList(),
-            ),
-            SamplesGalleryGroup(
-                name = "Advanced Samples",
-                samples = CollapsingTopBarSampleGroups.Advanced.toMutableStateList(),
-            ),
-            SamplesGalleryGroup(
-                name = "Playground Samples",
-                samples = CollapsingTopBarSampleGroups.Playground.toMutableStateList(),
-            ),
-        )
     }
 }
