+++ sample/shared/build.gradle.kts
@@ -37,6 +37,7 @@ kotlin {
             implementation(libs.compose.multiplatform.components.resources)
             implementation(libs.compose.multiplatform.material3)
             implementation(libs.compose.multiplatform.materialIconsCore)
+            implementation(libs.compose.multiplatform.navigationEvent)
         }
     }
 }
